pyproject.toml
[project]
name = "canary"
version = "0.1.0"
description = "ESG regulatory change monitoring agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "tenacity>=8",
    "beautifulsoup4>=4",
    "lxml>=5",
    "anthropic>=0.34",
    "langchain-anthropic>=0.3",
    "langgraph>=0.4",
    "langgraph-checkpoint-sqlite>=2",
    "pydantic>=2",
    "python-dotenv>=1",
    "pyyaml>=6",
]

[project.scripts]
canary = "canary.scheduler:main"

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.34",
    "ruff>=0.8",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: hits live external services",
    "llm: requires LLM API call",
]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/canary"]
